Tracking:  - Add the main screen structure into Navigation.kt.  - Make skeleton implementations of main 3 screens.  - Add GoogleMaps to the Tracking screen.  - Refactor the Room entities to have IDs as Longs instead of Ints.

+++ core-database/src/main/java/android/template/core/database/Trip.kt
@@ -11,7 +11,7 @@ import kotlinx.coroutines.flow.Flow
 @Entity
 data class Trip(
     @PrimaryKey(autoGenerate = true)
-    val id: Int = 0,
+    val id: Long = 0L,
     val startTimestamp: Long,
     val duration: Long,
     val distance: Double
@@ -31,5 +31,5 @@ interface TripDao {
     @Insert
     suspend fun insertTrip(
         item: Trip
-    ): Int
+    ): Long
 }
